+++ src/Win2Mqtt.Application/Win2MqttBackgroundService.cs
@@ -2,25 +2,43 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Win2Mqtt.Options;
+using Win2Mqtt.SystemMetrics;
 
 namespace Win2Mqtt.Application
 {
     public class Win2MqttBackgroundService(
-        Win2MqttService service,
+        IMqttConnectionManager connectionManager,
+        ISensorFactory sensorFactory,
+        IMessagePublisher publisher,
+        IMqttSubscriber subscriber,
         IOptionsMonitor<Win2MqttOptions> options,
         ILogger<Win2MqttBackgroundService> logger) : BackgroundService
     {
         private readonly static SemaphoreSlim _semaphore = new(1, 1);
+        private readonly IEnumerable<ISensorWrapper> _activeSensors = sensorFactory.GetEnabledSensors();
 
 
-        public override async Task StartAsync(CancellationToken cancellationToken)
+        public override async Task StartAsync(CancellationToken stoppingToken)
         {
+            // Connect to MQTT broker, Subscribe to incoming messages and Publish Home Assistant discovery messages/online status
+            // Connect to MQTT broker
+            await connectionManager.ConnectAsync(stoppingToken);
+
+            // Subscribe to incoming messages
+            await subscriber.SubscribeAsync(stoppingToken);
+
+            // Publish Home Assistant discovery messages
+            foreach (var sensor in _activeSensors)
+            {
+                await publisher.PublishSensorDiscoveryMessage(sensor.Metadata, stoppingToken);
+            }
+
+            // Publish online status
+            await publisher.PublishOnlineStatus(stoppingToken);
+
+            await base.StartAsync(stoppingToken);
             logger.LogInformation("Worker started");
 
-            // Connect to MQTT broker, Subscribe to incoming messages and Publish Home Assistant discovery messages/online status
-            await service.StartAsync(cancellationToken);
-
-            await base.StartAsync(cancellationToken);
         }
 
         protected override async Task ExecuteAsync(CancellationToken stoppingToken)
@@ -34,7 +52,11 @@ namespace Win2Mqtt.Application
                     try
                     {
 
-                        await service.CollectAndPublish(stoppingToken);
+                        foreach (var sensor in _activeSensors)
+                        {
+                            var sensorValue = await sensor.CollectAsync();
+                            await publisher.PublishSensorValue(sensor, sensorValue, stoppingToken);
+                        }
                     }
                     finally
                     {
@@ -67,7 +89,11 @@ namespace Win2Mqtt.Application
 
         public override async Task StopAsync(CancellationToken cancellationToken)
         {
-            await service.StopAsync(cancellationToken);
+            // Disonnect from MQTT broker
+            await publisher.PublishOfflineStatus(cancellationToken);
+
+            // Publish offline status
+            await connectionManager.DisconnectAsync(cancellationToken);
             await base.StopAsync(cancellationToken);
             logger.LogInformation("Worker stopped.");
         }
