+++ src/Win2Mqtt.LinuxService/Win2Mqtt.LinuxService.csproj
@@ -1,34 +1,32 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Worker">
 
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <OutputType>Exe</OutputType>
-        <PublishSingleFile Condition="'$(Configuration)' == 'Release'">true</PublishSingleFile>
-        <RuntimeIdentifier>linux-x64</RuntimeIdentifier>
-        <PlatformTarget>x64</PlatformTarget>
-    </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<OutputType>Exe</OutputType>
+		<PublishSingleFile Condition="'$(Configuration)' == 'Release'">true</PublishSingleFile>
+		<RuntimeIdentifier>linux-x64</RuntimeIdentifier>
+		<PlatformTarget>x64</PlatformTarget>
+	</PropertyGroup>
 
-    <PropertyGroup>
-        <!-- https://learn.microsoft.com/en-us/dotnet/core/extensions/workers#template-defaults -->
-        <ServerGarbageCollection>true</ServerGarbageCollection>
-    </PropertyGroup>
+	<PropertyGroup>
+		<!-- https://learn.microsoft.com/en-us/dotnet/core/extensions/workers#template-defaults -->
+		<ServerGarbageCollection>true</ServerGarbageCollection>
+	</PropertyGroup>
 
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="8.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Options.DataAnnotations" Version="8.0.0" />
-        <PackageReference Include="Serilog" Version="4.2.0" />
-        <PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
-        <PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
-        <PackageReference Include="Serilog.Settings.Configuration" Version="8.0.4" />
-        <PackageReference Include="Serilog.Sinks.File" Version="7.0.0" />
-    </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+		<PackageReference Include="Serilog" Version="4.2.0" />
+		<PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
+		<PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
+		<PackageReference Include="Serilog.Settings.Configuration" Version="8.0.4" />
+		<PackageReference Include="Serilog.Sinks.File" Version="7.0.0" />
+	</ItemGroup>
 
-    <ItemGroup>
-        <ProjectReference Include="..\Win2Mqtt.Common\Win2Mqtt.Common.csproj" />
-        <ProjectReference Include="..\Win2Mqtt.Application\Win2Mqtt.Application.csproj" />
-    </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\Win2Mqtt.Common\Win2Mqtt.Common.csproj" />
+		<ProjectReference Include="..\Win2Mqtt.Application\Win2Mqtt.Application.csproj" />
+	</ItemGroup>
 
 </Project>
