+++ src/Win2Mqtt.WindowsService/Program.cs
@@ -1,4 +1,3 @@
-using CliWrap;
 using Serilog;
 using Win2Mqtt;
 using Win2Mqtt.Application;
@@ -7,67 +6,39 @@ using Win2Mqtt.HomeAssistant;
 using Win2Mqtt.Options;
 using Win2Mqtt.SystemActions.Windows;
 using Win2Mqtt.SystemMetrics.Windows;
+using Win2Mqtt.WindowsService;
 
 Log.Logger = new LoggerConfiguration()
     .Enrich.FromLogContext()
     .WriteTo.Console()
     .CreateBootstrapLogger();
 
+// Manage installation and uninstallation of the service (CliWrap)
+if (await WindowsServiceInstaller.HandleServiceInstallationAsync(args))
+    return;
+
 try
 {
-    // Manage installation and uninstallation of the service (CliWrap)
-    var argsList = args.Select(a => a.ToLowerInvariant()).ToList();
-    if (argsList.Contains("/install"))
-    {
-        Log.Information("Install service.");
-        await Cli.Wrap("sc")
-            .WithArguments(new[] { "create", Constants.ServiceName, $"binPath=\"{Environment.ProcessPath}\"", "start=auto" })
-            .ExecuteAsync();
-        return;
-    }
-    else if (argsList.Contains("/uninstall"))
-    {
-        Log.Information("Uninstall service.");
-        await Cli.Wrap("sc")
-            .WithArguments(new[] { "stop", Constants.ServiceName })
-            .ExecuteAsync();
-
-        await Cli.Wrap("sc")
-            .WithArguments(new[] { "delete", Constants.ServiceName })
-            .ExecuteAsync();
-        return;
-    }
-
     Log.Information("Start application.");
 
     var builder = Host.CreateApplicationBuilder(args);
     builder.Services
+        .AddHostedService<Win2MqttBackgroundService>();
+
+    builder.Services
         .AddSerilog((services, loggerConfiguration) => loggerConfiguration
             .ReadFrom.Configuration(builder.Configuration)
             .ReadFrom.Services(services)
-            .Enrich.FromLogContext());
+            .Enrich.FromLogContext()
+            .WriteTo.Console());
 
     builder.Services
-        .AddOptionsWithValidateOnStart<Win2MqttOptions>()
-            .BindConfiguration(Win2MqttOptions.SectionName)
-            .ValidateDataAnnotations();
-
-    builder.Services
-        .PostConfigure<Win2MqttOptions>(o =>
-        {
-            o.Sensors = new Dictionary<string, SensorOptions>(o.Sensors, StringComparer.OrdinalIgnoreCase);
-            o.MultiSensors = new Dictionary<string, MultiSensorOptions>(o.MultiSensors, StringComparer.OrdinalIgnoreCase);
-            o.Listeners = new Dictionary<string, ListenerOptions>(o.Listeners, StringComparer.OrdinalIgnoreCase);
-        });
-
-    builder.Services
-        .AddWindowsService(options => options.ServiceName = $"{Constants.AppId} Service")
+        .AddWin2MqttOptions()
         .AddMqtt2NetBroker()
         .AddHomeAssistant()
         .AddWindowsSystemMetrics()
         .AddWindowsSystemActions()
-        .AddSingleton<Win2MqttService>()
-        .AddHostedService<Win2MqttBackgroundService>();
+        .AddWindowsService(options => options.ServiceName = $"{Constants.AppId} Service");
 
     await builder.Build().RunAsync();
 }
