+++ src/Win2Mqtt.WindowsService/appsettings.Development.json
@@ -2,11 +2,6 @@
     "Serilog": {
         "MinimumLevel": {
             "Default": "Debug"
-        },
-        "WriteTo": [
-            {
-                "Name": "Console"
-            }
-        ]
+        }
     }
 }
