+++ src/Win2Mqtt.WindowsService/appsettings.json
@@ -10,7 +10,7 @@
   },
   "Win2MQTT": {
     "Broker": {
-      "Server": "homeassistant.local",
+      "Server": "192.168.1.4",
       "Port": 1883,
       "Username": "win2mqtt",
       "Password": "win2mqtt"
